Added names and minor refactorings
Added typeNames using the CrestItemTypeRepository.  Might implement one that reads the sde yaml but that's a pretty massive yaml and I'm not interested in bloating my runtime memory usage.

Refactored some interfaces into the empty eve-sde project.  Might move more later but have some potential tangling to sort out.

+++ aba-industry/industry-common/src/main/java/com/aba/industry/service/IndustryCalculationService.java
@@ -26,14 +26,14 @@ import com.aba.industry.model.Decryptor;
 public interface IndustryCalculationService {
     /**
      * Convenience overload of
-     * {@link IndustryCalculationService#calculateBuildCosts(String, Long, IndustrySkillConfiguration, InventionSkillConfiguration, Integer, Integer, Decryptor, boolean, boolean)}
+     * {@link IndustryCalculationService#calculateBuildCosts(String, Integer, IndustrySkillConfiguration, InventionSkillConfiguration, Integer, Integer, Decryptor, boolean, boolean)}
      * which assumes all 5's for skills, true for findCurrentPrices and false for useBuildOrBuyConfigurations.
      *
      * @param systemName
      * @param outputTypeId
      * @return
      */
-    BuildCalculationResult calculateBuildCosts ( String systemName, Long outputTypeId );
+    BuildCalculationResult calculateBuildCosts ( String systemName, Integer outputTypeId );
 
     /**
      * Provides a mechanism to easily calculate the output of a full build with configuration options to either use
@@ -56,7 +56,7 @@ public interface IndustryCalculationService {
      */
     BuildCalculationResult calculateBuildCosts (
             String systemName,
-            Long outputTypeId,
+            Integer outputTypeId,
             IndustrySkillConfiguration industrySkills,
             InventionSkillConfiguration inventionSkills,
             Integer meLevel,
