+++ aba-industry/industry-fetcher/src/main/java/com/aba/industry/fetch/client/impl/StaticDataExportBlueprintYamlService.java
@@ -46,7 +46,8 @@ public class StaticDataExportBlueprintYamlService implements BuildRequirementsPr
     }
 
     @Override
-    public BlueprintData getBlueprintData ( final Long typeId ) {
+    public BlueprintData getBlueprintData ( final Integer typeId )
+    {
         BlueprintData result = new BlueprintData();
 
         Blueprint productionBp = null;
